これとどっちがいいのかな

+++ src/setNum.h
@@ -1,10 +1,11 @@
 short set(int pt,int  setNum, short array){
     short offset = setNum << 3*(pt - 1);
-    short bottomPadding = pow(2, (3 * setNum - 2));
-    short headPadding = pow(2, (sizeof(array) * (8 - pt * 3) - 1));
+    short bottomPadding = (short)pow(2.0, ((float)(3 * setNum - 2)));
+    short headPadding = (short)pow(2.0, ((float)(sizeof(array) * (8 - pt * 3) - 1)));
     short newArray = offset + bottomPadding + headPadding;
 }
 
+/*
 int pow(int p, int n){
     int powed;
     for(int i = 0; i < n; i++){
@@ -12,3 +13,4 @@ int pow(int p, int n){
     }
     return powed;
 }
+*/
